backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
stripe==6.6.0
requests==2.31.0
email-validator==2.0.0 